+++ noodles-bam/src/reader/record/position.rs
@@ -15,7 +15,8 @@ where
 
     match src.get_i32_le() {
         MISSING => Ok(None),
-        n => usize::try_from(n + 1)
+        n => usize::try_from(n)
+            .map(|m| m + 1)
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
             .map(Position::new),
     }
@@ -48,6 +49,16 @@ mod tests {
             Err(e) if e.kind() == io::ErrorKind::InvalidData,
         ));
 
+        Ok(())
+    }
+
+    #[cfg(not(target_pointer_width = "16"))]
+    #[test]
+    fn test_get_position_with_max_position() -> Result<(), Box<dyn std::error::Error>> {
+        let data = i32::MAX.to_le_bytes();
+        let mut src = &data[..];
+        let expected = Position::try_from(1 << 31)?;
+        assert_eq!(get_position(&mut src)?, Some(expected));
         Ok(())
     }
 }